Fix tests broken by adapter move
